migrating OS detection, OS enums and usesCtrlKey functions to their own file in common/src/utils
